// src/app/users/models/user.ts
export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
        lat: string;
        lng: string;
    };
}

export interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
}

export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address?: Address;  // Теперь это объект
    phone?: string;
    website?: string;
    company?: Company;  // Теперь это объект
}